rename openApiDoc folder to match ADR
