+++ front/src/app/pages/open-api-doc/OpenApiDocPage.tsx
